requirements.txt
requests
Pillow==10.3.0
